requirements.txt
# Add required dependencies
openai
requests
python-dotenv
beautifulsoup4